+++ packages/elements/src/internals/machines/sign-in/types/verification.types.ts
@@ -18,11 +18,17 @@ export type SignInVerificationTags =
 
 export type SignInVerificationSubmitEvent = { type: 'SUBMIT' };
 export type SignInVerificationFactorUpdateEvent = { type: 'STRATEGY.UPDATE'; factor: SignInFactor | undefined };
+export type SignInVerificationRetryEvent = { type: 'RETRY' };
+export type SignInVerificationStrategyRegisterEvent = { type: 'STRATEGY.REGISTER'; factor: SignInFactor };
+export type SignInVerificationStrategyUnregisterEvent = { type: 'STRATEGY.UNREGISTER'; factor: SignInFactor };
 
 export type SignInVerificationEvents =
   | ErrorActorEvent
   | SignInVerificationSubmitEvent
-  | SignInVerificationFactorUpdateEvent;
+  | SignInVerificationFactorUpdateEvent
+  | SignInVerificationRetryEvent
+  | SignInVerificationStrategyRegisterEvent
+  | SignInVerificationStrategyUnregisterEvent;
 
 // ---------------------------------- Input ---------------------------------- //
 
@@ -39,13 +45,25 @@ export interface SignInVerificationContext {
   formRef: ActorRefFrom<typeof FormMachine>;
   parent: ActorRefFrom<TSignInRouterMachine>;
   loadingStep: 'verifications';
+  registeredStrategies: Set<SignInFactor>;
+  resendable: boolean;
+  resendableAfter: number;
 }
+
+// ---------------------------------- Delays ---------------------------------- //
+
+export const SignInVerificationDelays = {
+  resendableTimeout: 1_000, // 1 second
+} as const;
+
+export type SignInVerificationDelays = keyof typeof SignInVerificationDelays;
 
 // ---------------------------------- Schema ---------------------------------- //
 
 export interface SignInVerificationSchema {
   context: SignInVerificationContext;
   input: SignInVerificationInput;
+  delays: SignInVerificationDelays;
   events: SignInVerificationEvents;
   tags: SignInVerificationTags;
 }