+++ packages/localizations/src/th-TH.ts
@@ -84,6 +84,7 @@ export const thTH: LocalizationResource = {
     'กรอกที่อยู่อีเมลภายใต้โดเมนนี้เพื่อรับรหัสและตรวจสอบโดเมน',
   formFieldLabel__organizationName: 'ชื่อ',
   formFieldLabel__organizationSlug: 'Slug',
+  formFieldLabel__passkeyName: undefined,
   formFieldLabel__password: 'รหัสผ่าน',
   formFieldLabel__phoneNumber: 'หมายเลขโทรศัพท์',
   formFieldLabel__role: 'บทบาท',
@@ -486,7 +487,7 @@ export const thTH: LocalizationResource = {
     passkey_not_supported: 'อุปกรณ์นี้ไม่รองรับพาสคีย์',
     passkey_registration_cancelled: 'การลงทะเบียนพาสคีย์ถูกยกเลิกหรือหมดเวลา',
     passkey_retrieval_cancelled: 'การยืนยันพาสคีย์ถูกยกเลิกหรือหมดเวลา',
-    passkeys_pa_not_supported: 'การลงทะเบียนต้องใช้ระบบยืนยันตัวตนของแพลตฟอร์ม แต่อุปกรณ์ไม่รองรับ',
+    passkey_pa_not_supported: 'การลงทะเบียนต้องใช้ระบบยืนยันตัวตนของแพลตฟอร์ม แต่อุปกรณ์ไม่รองรับ',
     passwordComplexity: {
       maximumLength: 'น้อยกว่า {{length}} ตัวอักษร',
       minimumLength: '{{length}} ตัวอักษรหรือมากกว่า',
@@ -670,6 +671,14 @@ export const thTH: LocalizationResource = {
       security: 'ความปลอดภัย',
       title: 'บัญชี',
     },
+    passkeyScreen: {
+      removeResource: {
+        messageLine1: undefined,
+        title: undefined,
+      },
+      subtitle__rename: undefined,
+      title__rename: undefined,
+    },
     passwordPage: {
       checkboxInfoText__signOutOfOtherSessions: 'ขอแนะนำให้ลงชื่อออกจากอุปกรณ์อื่น ๆ ที่อาจใช้รหัสผ่านเก่าของคุณ',
       readonly:
@@ -755,6 +764,11 @@ export const thTH: LocalizationResource = {
           destructiveActionTitle: 'ลบ',
           headerTitle: 'แอปพลิเคชันยืนยันตัวตน',
         },
+      },
+      passkeysSection: {
+        menuAction__destructive: undefined,
+        menuAction__rename: undefined,
+        title: undefined,
       },
       passwordSection: {
         primaryButton__setPassword: 'ตั้งรหัสผ่าน',